+++ tests/test_main.py
@@ -1,6 +1,14 @@
+import os
+
 import pytest
 
-from isce2_topsapp.__main__ import ensure_earthdata_credentials, esd_threshold_argument, true_false_string_argument
+from isce2_topsapp.__main__ import (
+    ESA_HOST,
+    check_esa_credentials,
+    ensure_earthdata_credentials,
+    esd_threshold_argument,
+    true_false_string_argument,
+)
 
 
 def test_main_check_earthdata_credentials_prefer_netrc(tmp_path, monkeypatch):
@@ -73,6 +81,65 @@ def test_main_check_earthdata_credentials_env_variables(tmp_path, monkeypatch):
     assert netrc.read_text() == 'machine foobar.nasa.gov login fizz password baz'
 
 
+def test_check_esa_credentials_params(tmp_path, monkeypatch):
+    with monkeypatch.context() as m:
+        m.setenv('ESA_USERNAME', 'env_username')
+        m.setenv('ESA_PASSWORD', 'env_password')
+        m.setenv('HOME', str(tmp_path))
+        (tmp_path / '.netrc').write_text(f'machine {ESA_HOST} login netrc_username password netrc_password')
+
+        check_esa_credentials('foo', 'bar')
+
+        assert os.environ['ESA_USERNAME'] == 'foo'
+        assert os.environ['ESA_PASSWORD'] == 'bar'
+
+
+def test_check_esa_credentials_env(tmp_path, monkeypatch):
+    with monkeypatch.context() as m:
+        m.setenv('ESA_USERNAME', 'foo')
+        m.setenv('ESA_PASSWORD', 'bar')
+        m.setenv('HOME', str(tmp_path))
+        (tmp_path / '.netrc').write_text(f'machine {ESA_HOST} login netrc_username password netrc_password')
+
+        check_esa_credentials(None, None)
+
+        assert os.environ['ESA_USERNAME'] == 'foo'
+        assert os.environ['ESA_PASSWORD'] == 'bar'
+
+
+def test_check_esa_credentials_netrc(tmp_path, monkeypatch):
+    with monkeypatch.context() as m:
+        m.delenv('ESA_USERNAME', raising=False)
+        m.delenv('ESA_PASSWORD', raising=False)
+        m.setenv('HOME', str(tmp_path))
+        (tmp_path / '.netrc').write_text(f'machine {ESA_HOST} login foo password bar')
+
+        check_esa_credentials(None, None)
+
+        assert os.environ['ESA_USERNAME'] == 'foo'
+        assert os.environ['ESA_PASSWORD'] == 'bar'
+
+
+def test_check_esa_credentials_missing(tmp_path, monkeypatch):
+    with monkeypatch.context() as m:
+        m.delenv('ESA_USERNAME', raising=False)
+        m.setenv('ESA_PASSWORD', 'env_password')
+        m.setenv('HOME', str(tmp_path))
+        (tmp_path / '.netrc').write_text('')
+        msg = 'Please provide.*'
+        with pytest.raises(ValueError, match=msg):
+            check_esa_credentials(None, None)
+
+    with monkeypatch.context() as m:
+        m.setenv('ESA_USERNAME', 'env_username')
+        m.delenv('ESA_PASSWORD', raising=False)
+        m.setenv('HOME', str(tmp_path))
+        (tmp_path / '.netrc').write_text('')
+        msg = 'Please provide.*'
+        with pytest.raises(ValueError, match=msg):
+            check_esa_credentials(None, None)
+
+
 def test_true_false_string_argument():
     assert true_false_string_argument('true') is True
     assert true_false_string_argument('TRUE') is True
@@ -90,11 +157,11 @@ def test_true_false_string_argument():
 
 
 def test_esd_threshold_argument():
-    assert esd_threshold_argument('-1') == -1.
-    assert esd_threshold_argument('0') == 0.
+    assert esd_threshold_argument('-1') == -1.0
+    assert esd_threshold_argument('0') == 0.0
     assert esd_threshold_argument('.25') == 0.25
     assert esd_threshold_argument('0.5') == 0.5
-    assert esd_threshold_argument('1') == 1.
+    assert esd_threshold_argument('1') == 1.0
 
     with pytest.raises(ValueError):
         esd_threshold_argument('-1.1')